Desks: Allow restoring desk root tasks on user switch
Prior logic to save visible root tasks on user-switch to restore to when
switching back did not account for non-leaves not being associated to a
particular user, so this change removes all user checks because
visibility is already only possible for the current user.

This fixes an issue with desktop windowing's desk roots and children not
being restored.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 404591902
Test: with a desktop open, switch users, switch back, check desktop
tasks are restored

Change-Id: I2d5883a274f8ec09778e9f7d36e08741a077034c