WindowDecor: Make exclusion region display-specific.
Exclusion region used to be just on a per-Shell-init instance, and we
drop anything that is non-primary displayId. This creates a mapping of
exclusion regions for each display, add/remove as see fit, and update
the window decorations on the displays accordingly.

Bug: 414786783
Test: atest DesktopModeWindowDecorViewModelTests
Flag: com.android.window.flags.enable_bug_fixes_for_secondary_display
Change-Id: I35e8cec3f0a8e312ac4113a42fb7296ff2d95f78