AnimatedBoundsLayoutListener jump-cuts rotations
This OnLayoutChangeListener animates smoothly between old and new
onscreen bounds of the View, but this only works well if its frame of
reference is not changing underneath it. When the parent bounds are
changing due to rotation, resize, or translation, detect that and
leave the View bounds alone.

Test: manual (open Internet settings, rotate screen 3x)
Flag: EXEMPT bugfix
Fix: 267689502
Change-Id: I383bd6a94128bc3542a914040b1dafe259058b27