Remove PredrawCallback when local ValueAnimator is complete
Instead of removing the PredrawCallback when the transition at large
has ended, we additionally remove it when the local ValueAnimator is
ended. This may help in certain situations where onTransitionEnd
isn't being fired as expected. This fix is safe, but I'm uncertain
that it'll solve the leak for all edge cases.

Bug: 414596090
Flag: NONE Bugfix
Test: Manually checked clock switch animation in several scenarios
Change-Id: I97fa223f21a677a931eb676522fff0ea751cc41c