Power: Add args for sleep/wakeup cmd to disable wakelock
PowerManagerService has a new flag to force wakelocks
to be disabled and we should be able to test it with
sleep/wakeup cmd as it will be a common suspend journey.

Add --disable-wakelocks arg to sleep cmd and
--restore-wakelocks arg to wakeup cmd for that. Former
will set the force flag to true before suspending and
latter will set the flag to false after resuming. It's
named as restore as it does not necessarily re-enable
wakelocks that are disabled by other features/flags.

Bug: 410932789
Test: Manually acquire a wakelock & verify these
      cmd power sleep --disable-wakelocks
      cmd power wakeup --restore-wakelocks
      cmd power wakeup 10000 --restore-wakelocks
Flag: EXEMPT test
Change-Id: Ia94aee3412380042b26d80d57bd2b660cfa12521