Default managed profiles to off on KEY_TYPE adjustments only
Because KEY_TYPE (classification) is the only adjustment (for now) for which we want the work profile to default to off if the user hasn't explicitly turned it on, store a list of users whose preferences for that adjustment have ever changed. Then if the relevant user is a managed profile, we include KEY_TYPE in the denied adjustments if it wasn't explicitly set one way or another by the user.

Also add some more test setup in NotificationAssistantsTest so we don't get different results about what user is a profile user depending on which mock (NMS, UserManager, UserManagerInternal...) we ask.

Fixes: 415768865
Test: NotificationAssistantsTest, manual via behavior + checking policy xml
Flag: android.app.notification_classification_ui
Change-Id: Ie2e27503cc42a451d954eb66703b01628130a213