[dual sim] use Image so we can properly set the size
StackedMobileIcon was using an Icon before, which seems not to set the
size of the vector image based on the passed-in modifier, and doesn't
have a parameter for ContentScale.

Image provides the things we need to ensure that the stacked mobile icon
is the same size as the single-SIM icon.

Test: manual
Fixes: 416762981
Flag: com.android.settingslib.flags.new_status_bar_icons
Flag: com.android.systemui.status_bar_root_modernization
Change-Id: I2ec4ca6b939e059ce6721adc5e6518410e08aa05