Use slowSpatialSpec() when flinging and animating to targetOffset
This CL changes the spring used to animate to the target offset when
flinging a STL transition, from defaultSpatialSpec() to
slowSpatialSpec(). The slow spec is better for large offsets, which is
the case for our main use case of STL (Flexiglass).

In the future, we could either make this a parameter of our transition
specs or, even better, decide that automatically based on the size of
the STL or the gesture distance.

Bug: 417444347
Test: Manual, observed the slower animation when releasing a swipe
 transition
Flag: com.android.systemui.scene_container
Change-Id: I12876009a3a6a0d5e340ca836a6773a4f5960321