Clear task bounds when ActivityOptions request 0,0
Support clearing multi-window task bounds if they are set to 0,0 in
launch bounds in ActivityOptions.
Previously, empty bounds in LaunchParams were treated as the default
value, and not applied.
This change introduces a new property in LaunchParams that is set by
TaskLaunchParamsModifier to indicate that the bounds are set.
This allows LaunchParamsController#layoutTask to determine that new
bounds are requested, even if they are 0,0.

Fixes a scenario with split screen where a task that previously had
bounds set, is launched to split screen. Split screen expects the task
to not have bounds set and that the task should inherit the bounds from
the root task. Allowing split to request 0,0 launch bounds through
ActivityOptions fixes the issue with task having bounds set on it.

Bug: 396315875
Test: atest WmTests:LaunchParamsControllerTests
Flag: EXEMPT, bugfix
Change-Id: Idc16473e488d7e4bfda98f8d49e2f6ae68762c97