Allow dreaming when charge limit is active
When a charge limit is active and the battery reaches the level of the
limit, the device will stop charging entirely. We still want to dream
in these cases as battery usage is not an issue since the device will
start charging again if the battery level goes below the limit.

Bug: 397717022
Fixes: 397717022
Test: atest DreamManagerServiceTest
      also manually verified on phone
Flag: android.service.dreams.allow_dream_with_charge_limit
Change-Id: Id5d67e6f687335bccb9edfe3787412f8c59d8c86