[Flexiglass] Don't allow scene changes by swipes while expanding notifs
When the user was dragging down on a notification in SingleShade,
and the stack was scrolled all the way to the top, at the end of
the expansion gesture the scene stack transitioned from Shade to
QuickSettings. The cure is to:
 - Dispatch touches to the scene framework, while it is expanding a
notification.
 - Create a nested scroller, that consumes all the drag amount while
the gesture is being used to expand a notification.
 - Rely on Modifier.disableSwipesWhenScrolling() to disable the scene
   change, as the previously added scroller has comsumed some of this
   gesture.
 - Cleanup the old mechanism of isCurrentGestureOverscroll(), as it
   was broken since an API was removed from the scene stack.

Bug: 416342077
Test: atest PlatformScenarioTests:android.platform.test.scenario.sysui.notification.DraggingToExpandGroupNotifications
Flag: com.android.systemui.scene_container

Change-Id: Idabdbf63dff7478e3d822236d1e2f789c9917df3