Re-design of squeeze effect haptics.
For better synchrony and the ability to play haptics on the LPP event,
we divide the complete effect into parts: An initial rumble, an LPP
indicator, and a zoomout effect. TopLevelWindowEffects play all of these
independently. The initial rumble is optional and can be enabled via a
System property.

Test: TopLevelWindowEffectsTest
Test: Verified haptics play correctly with and without the initial
  rumble
Flag: com.android.systemui.shared.enable_lpp_assist_invocation_haptic_effect
Bug: 417200685
Bug: 417496751
Change-Id: I1040b2862acc95ef87306f8a86c186c5a8a91a73