Add a software rate-limiter to LockSettingsService
Add a SoftwareRateLimiter and wire it up to LockSettingsService.  The
SoftwareRateLimiter implements duplicate wrong guess detection and
implements a rate-limiting schedule that is more secure than the one
that Android has traditionally used.  It complements the "hardware"
rate-limiter and allows faster testing and deployment of a new
rate-limiting schedule before hardware is updated which generally takes
longer.  For more details, see the SoftwareRateLimiter class comment.

Bug: 395976735
Test: atest FrameworksServicesTests:com.android.server.locksettings
Test: manual test with the lock screen on phone and watch
Flag: android.security.software_ratelimiter
Change-Id: I79dd3595995603f53e62f714de6e1c46b46d9d87