[flexiglass] Add back button on bouncer
This CL adds a button to the bouncer (for all layouts) in the lower left corner. When clicked, it will navigate back to keyguard.

The button is guarded by its a config parameter (config_improveLargeScreenInteractionOnLockscreen) and only added to the new flexiglass version of the bouncer.

Fixes: 418724618
Flag: com.android.systemui.scene_container
Test: atest BouncerOverlayContentViewModelTest
Test: atest BouncerContentComposeTest

Change-Id: I98704b7278e4a4861bfdb560b0d444f8caee16cd