Make weaverVerify() return a WeaverReadResponse
Currently weaverVerify() translates the WeaverReadResponse into a
VerifyCredentialResponse, including stashing the Weaver secret in
VerifyCredentialResponse#mGatekeeperHAT (despite the name of this field
not matching this use).  In some cases the VerifyCredentialResponse from
weaverVerify() gets used as the final VerifyCredentialResponse, whereas
in other cases the caller just uses it to check the result of the Weaver
operation and then builds the real VerifyCredentialResponse later.

This makes things unnecessarily convoluted.  Instead, make
weaverVerify() just return the WeaverReadResponse and have the callers
handle it appropriately.  No functional change.

Test: atest FrameworksServicesTests:com.android.server.locksettings
Bug: 320392352
Bug: 395976735
Flag: EXEMPT refactor
Change-Id: I965bfc36e8d3f68223819feb8ac29ca933be6668