Make ShellTestCase try to delete TestableLooper
TestableLooper is dangerous due to the ease of adding it with an
annotation and the poor explanation of what the implications of doing
this are (it is REQUIRED to manually clean up in an @After block to
avoid memory leaks which break large test suites).

The majority of tests in wm/shell are inheriting from ShellTestCase so
we can claw some memory back by having the superclass try to do the
right thing even if individual tests forget. This should become a
warning and eventually an error, and eventually a full ban.

Test: atest WMShellUnitTests
Bug: 261039202
Bug: 419430883
Flag: EXEMPT bugfix
Change-Id: Ie40c897d1cd6a86f8013b6c4372797d31482ed59