Use user context when loading pip menu icons
This is a counterpart of ag/28640753, but for pip2

There is a bug involving trying to load app icons when using secondary
users (or HSUM). Since we load icons using the systemui context which
is always user 0, we might not be able to load the icons if the app
is not installed on user 0.

Fixed by using the current user context when loading menu icons.

Flag: com.android.wm.shell.enable_pip2
Bug: 419687679
Test: Check PiP menu in secondary user and/or HSUM
Change-Id: I33aa241ce44220b5132ce7b3d65c4617a4245bdd