[3/n] Add AppCompatTransitionInfo to ActivityTransitionInfo
The AppCompatTransitionInfo encapsulates information related
to App Compat in case the activity, target of a transition,
is letterboxed.
At the moment AppCompatTransitionInfo contains the letterboxBounds
which would not be available otherwise because of the missing
TaskInfo in the Transition's Change.

The AppCompatTransitionInfo is optional property of the existing
ActivityTransitionInfo.

Flag: EXEMPT bug fixing
Bug: 377875664
Test: atest FrameworksCoreTests:TransitionInfoTest
Test: atest WmTests:AppCompatUtilsTest
Test: atest WmTests:TransitionTests

Change-Id: I4f51148bd68e0afdd9257b670a03a6c354892688