Remove Nothing return from onActivated()
The `activate()` function returns Nothing as it is the API surface and
it communicates to the user that the function never returns.
`onActivated()` in turn is only called internally and is supposed to
be overriden. Returning Nothing forces the overwrite to add
`awaitCancellation` which is unnecessary because it is ok to make it
return early.

Bug: b/420591935
Test: Refactor only
Flag: com.android.systemui.scene_container
Change-Id: Ie1541f303e4930f36775b8e88d3b43dfbde6d95e