Add ANR Specific Subreasons to AppExitInfo
This update introduces specific ANR subreasons to the
ApplicationExitInfo subreason list. Additionally, TimeoutRecord is
modified to correlate its TimeoutKind with a corresponding
ApplicationExitInfo ANR subreason. The subreason is then passed to
ApplicationExitInfo objects when the system identifies an unresponsive
app to be used by apps for further analysis.

This change does not populate the ANR subreason when the user kills the
app from the ANR dialog.

Bug: 295599954
Test: atest CtsAppExitTestCases:ActivityManagerAppExitInfoTest
Flag: android.app.include_anr_subreason
Change-Id: I9e427ec74f7e149f6de050b5d3747de913e75ebc