Use stable bounds orientation for initial bounds calculation
Using the existing task orientation is not a reliable way to calculate
the required initial bounds as there is a possibility the destination
orientation might be different to the current task one. This can be
the case with split screen to desk as well as moving tasks to external
displays.

Instead use the stable bounds orientation of the destination display.

Flag: com.android.window.flags.enable_windowing_dynamic_initial_bounds
Test: atest WMShellUnitTests:DesktopTasksControllerTest
Fixes: 404887079
Change-Id: Ic013146710bd2aa1f8afef32077f157312b1ccf6