Stop depending on transitions for task limit tile handling.
A bug was introduced when changing the transition for minimizing a task
when the task limit is reached from TRANSIT_OPEN to
TRANSIT_DESKTOP_MODE_TASK_LIMIT_MINIMIZE which broke the signal tiling was
expecting to handle this minimize event.

This CL abandons depending on transitions for this event and migrates the
signaling to DesktopTasksLimiter to avoid future changes breaking tiling.

Flag: com.android.window.flags.enable_tile_resizing
Test: Unit tests and on device testing
Bug: 419445881
Change-Id: I5f76eafb60ecdfc6fe78aa4a04a5cb59d42018fd