Check resolution changes in DisplayInfo.equals.
The previous check for display updates only checked if the display's
id changed when compareOnlyBasicChanges=false. With
compareOnlyBasicChanges=true
this could cause updates to be ignored.

This change adds a check for the current active display mode in
DisplayInfo.equals, and only returns that the info is equal if the
display mode ids are the same, or both mode's dimensions are the same.

Test: atest DisplayInfoTest
Test: manual see change is received in Settings
Bug: 413014130
Bug: 424418347
Flag: EXEMPT simple bugfix
Change-Id: I9c103002a1a99f3921546238c1612869310046fe